% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audio_description.R
\name{get_audio_description}
\alias{get_audio_description}
\title{Get audio from text string}
\usage{
get_audio_description(
  text,
  dest_file,
  pcdas_token = NULL,
  throttle_rate = 1,
  max_tries = 10
)
}
\arguments{
\item{text}{character. A text string.}

\item{dest_file}{character. Path to destination audio file. The output is a mp3 binary.}

\item{pcdas_token}{character. PCDaS API token. If not provided, the function will look for it on renvirom.}

\item{throttle_rate}{Rate of requests per second allowed. Defaults to 1 request per second.}

\item{max_tries}{Max number of retries before fail. Defaults to 10.}
}
\value{
an audio file.
}
\description{
Get audio from text string
}
\examples{
file <- tempfile(fileext = ".mp3")
get_audio_description(text = "Ciência de Dados, para nós da PCDaS, é um campo de estudo que se destaca pela capacidade de auxiliar a descoberta de informação útil a partir de grandes ou complexas bases de dados.", dest_file = file)
}
