% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcdas_token_renviron.R
\name{get_pcdas_token_renviron}
\alias{get_pcdas_token_renviron}
\title{Get PCDaS token from renviron}
\usage{
get_pcdas_token_renviron()
}
\value{
A character.
}
\description{
Get PCDaS token from renviron
}
\details{
PCDaS token may be stored at renviron with \code{Sys.setenv("pcdas_token" = "YOUR_PCDAS_TOKEN")}. This function retrivies the token if stored at renviron.
}
