% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinasc.R
\name{get_sinasc}
\alias{get_sinasc}
\title{Get new births data from SINASC}
\usage{
get_sinasc(
  agg,
  agg_time = "year",
  ano,
  pcdas_token = NULL,
  sexo = NULL,
  more_filters = NULL,
  fetch_size = 65000
)
}
\arguments{
\item{agg}{character. Spatial aggregation level. \code{uf_res} for UF of residence. \code{uf_ocor} for UF of occurrence. \code{regsaude_res} for regiao de saude of residence. \code{regsaude_ocor} for regiao de saúde of occurence. \code{mun_res} for municipality of residence. \code{mun_ocor} for municipality of ocurrence.}

\item{agg_time}{character. Time aggregation level. \code{year} for yearly data. \code{month} for monthly data. \code{week} for weekly data. Defaults to \code{year}.}

\item{ano}{numeric. Year of birth}

\item{pcdas_token}{character. PCDaS API token. If not provided, the function will look for it on renvirom.}

\item{sexo}{character. Sex of the new birth \code{Masculino} for males, \code{Feminino} for females and \code{Ignorado} for unknown.}

\item{more_filters}{character. Additional filters can be added by using this parameter, with a SQL query.}

\item{fetch_size}{character. Pagination size for API call.}
}
\description{
Retrieves new births data from PCDaS API.
}
\details{
This function uses raw data from the Sistema de Informações de Nascidos Vivos (SINASC) available at the PCDaS API. A documentation about this data can be found at \url{https://pcdas.icict.fiocruz.br/conjunto-de-dados/sistema-de-informacao-sobre-nascidos-vivos/}.
}
\examples{
# Some examples
get_sinasc(agg = "mun_res", ano = 2010)
get_sinasc(agg = "uf_ocor", ano = 2010)

}
