% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text_description.R
\name{get_text_description}
\alias{get_text_description}
\title{Get a description of a data frame}
\usage{
get_text_description(
  df,
  prompt,
  pcdas_token = NULL,
  throttle_rate = 1,
  max_tries = 10
)
}
\arguments{
\item{df}{a data frame.}

\item{prompt}{character. A prompt asking how the data should be described.}

\item{pcdas_token}{character. PCDaS API token. If not provided, the function will look for it on renvirom.}

\item{throttle_rate}{Rate of requests per second allowed. Defaults to 1 request per second.}

\item{max_tries}{Max number of retries before fail. Defaults to 10.}
}
\value{
a character string.
}
\description{
Get a description of a data frame
}
\examples{
get_text_description(df = iris, prompt = "This json data contains measurements of sepal length and width and petal length and width. Write a small paragraph describing the species. Do not mention the file format.")

}
