% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sih.R
\name{get_sih}
\alias{get_sih}
\title{Get hospitalization data from SIH}
\usage{
get_sih(
  agg,
  ano,
  pcdas_token = NULL,
  sexo = NULL,
  idade_a = NULL,
  idade_b = NULL,
  more_filters = NULL
)
}
\arguments{
\item{agg}{character. Spatial aggregation level. \code{uf_res} for UF of residence. \code{uf_ocor} for UF of occurrence. \code{regsaude_res} for regiao de saude of residence. \code{regsaude_ocor} for regiao de saúde of occurence. \code{mun_res} for municipality of residence. \code{mun_ocor} for municipality of ocurrence.}

\item{ano}{numeric. Year of birth}

\item{pcdas_token}{character. PCDaS API token. If not provided, the function will look for it on renvirom.}

\item{sexo}{character. Sex of the new birth \code{Masculino} for males, \code{Feminino} for females and \code{Ignorado} for unknown.}

\item{idade_a}{numeric. Minimum age of the patient, in years.}

\item{idade_b}{numeric. Maximum age of the patient, in years.}

\item{more_filters}{character. Additional filters can be added by using this parameter, with a SQL query.}
}
\description{
Retrieves hospitalization data from PCDaS API.
}
\details{
This function uses raw data from the Sistema de Informações de Hospitalares do SUS (SIH) available at the PCDaS API. A documentation about this data can be found at \url{https://pcdas.icict.fiocruz.br/conjunto-de-dados/sistema-de-informacoes-hospitalares-do-sus-sihsus/}.

If \code{idade_a} is supplied, the query will filter records with age less or equal. If \code{idade_b} is supplied, the query will filter records with age more or equal. If both are supplied, the query will filter records in the interval, closed on both sides.
}
\examples{
# Some examples
get_sih(agg = "mun_res", ano = 2010)
get_sih(agg = "uf_ocor", ano = 2010)

}
